Make an IDE friendly file for Step Definition "definedParameterTypes" ...
If you want to generate the hints file, run "npm run generate:hints" ... this will allow our IDE to know what parameter matches we have for step definitions.

Otherwise, because we are generating the parameter Types dynamically, it is completely clueless as to how you might match step definitions and just assumes we did not define things properly.